avr port

+++ src/blob_json.c
@@ -239,10 +239,10 @@ static void blob_format_element(struct strbuf *s, struct blob_field *attr, bool 
 		sprintf(buf, "%d", be16_to_cpu(*(uint16_t *)data));
 		break;
 	case BLOB_FIELD_INT32:
-		sprintf(buf, "%lu", (int32_t) be32_to_cpu(*(uint32_t *)data));
+		sprintf(buf, "%d", (int) be32_to_cpu(*(uint32_t *)data));
 		break;
 	case BLOB_FIELD_INT64:
-		sprintf(buf, "%llu", (int64_t) be64_to_cpu(*(uint64_t *)data));
+		sprintf(buf, "%lld", (long long int) be64_to_cpu(*(uint64_t *)data));
 		break;
 	case BLOB_FIELD_FLOAT32: 
 		sprintf(buf, "%f", (double) unpack754_32(be32_to_cpu(*(uint32_t*)data)));
